Rename user detail folder
