+++ src/app/user-detail-viewer/profile-icon/profile-icon.component.css
