+++ src/app/user-detail-viewer/user-detail-viewer.component.ts
